backend/requirements-grading-agent.txt
# Quality Grading Agent Requirements
# FastAPI Backend with YOLO + EfficientNet-B0

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Deep Learning
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.220  # YOLO v8 segmentation
timm==0.9.12  # EfficientNet models

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Optional: For production deployment
# gunicorn==21.2.0
# prometheus-client==0.19.0
